+++ CHANGES.md
@@ -225,3 +225,6 @@
 ### v2025.05.08
 - Phonak: fix Roger Upgrader update title for documents
 - Add `Quick Scan` option to get all major hearing aids software update across several market at once (ONLY available in Turbo Mode, Extra Menu)
+
+### Pre-release (work-in-progress)
+- Underhood bug fix